pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipeline-coach"
version = "0.1.0"
description = "Daily pipeline hygiene coach for Twenty CRM"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "langgraph>=0.2.0",
    "dspy>=3.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "resend>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
]
mcp = ["mcp>=1.0.0"]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
